refactor(contract): move TEE attestation lifecycle into api/attestation.rs
